requirements.txt
fastapi[standard]
uvicorn
pydantic
mysql-connector-python
mysql
python-dotenv
python-multipart
aiosmtplib
python-jose
pydantic[email]